:recycle: getMyRequests
