+++ src/client/mystatus/myrequests.rs
@@ -1,38 +1,14 @@
+use crate::client::HEADER_KEY;
 use crate::query::query_string;
-use anyhow::anyhow;
-use serde::{Deserialize, Serialize};
+use crate::response::error::{ErrorResponse, ErrorResponseBody};
+use crate::response::mystatus::myrequests::{GetMyRequestsResponse, GetMyRequestsResponseBody};
 use std::collections::HashMap;
 
 const RESOURCE: &str = "myrequests";
-const HEADER_KEY: &str = "X-Collaboflow-Authorization";
 
 pub struct MyRequests {
     url: String,
     authorization_header: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct GetMyRequestsResponse {
-    pub app_cd: i32,
-    pub offset: i32,
-    pub limit: i32,
-    pub total_count: i32,
-    pub decision_count: i32,
-    pub error: bool,
-    pub records: Vec<MyRequestRecord>,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct MyRequestRecord {
-    pub id: i32,
-    pub document_id: i32,
-    pub document_number: String,
-    pub title: String,
-    pub request_usercd: String,
-    pub request_name: String,
-    pub request_date: String,
-    pub flow_status: String,
-    pub link: String,
 }
 
 impl MyRequests {
@@ -46,19 +22,57 @@ impl MyRequests {
     pub async fn get(
         &self,
         query_params: HashMap<String, String>,
-    ) -> anyhow::Result<GetMyRequestsResponse> {
+    ) -> Result<GetMyRequestsResponse, ErrorResponse> {
         let request_url = format!("{}?{}", &self.url, query_string(query_params));
 
         let http_client = reqwest::Client::new();
-        let resp = http_client
+        let result = http_client
             .get(request_url)
             .header(HEADER_KEY, &self.authorization_header)
             .send()
-            .await?;
+            .await;
 
-        match resp.json::<GetMyRequestsResponse>().await {
-            Ok(myrequests) => Ok(myrequests),
-            Err(err) => Err(anyhow!(err)),
+        match result {
+            Ok(resp) => {
+                let status = resp.status().as_u16();
+
+                if status == 200 {
+                    match resp.json::<GetMyRequestsResponseBody>().await {
+                        Ok(body) => Ok(GetMyRequestsResponse { status, body }),
+                        Err(err) => {
+                            let body = ErrorResponseBody {
+                                error: true,
+                                messages: vec![err.to_string()],
+                            };
+                            let error_response = ErrorResponse { status, body };
+                            Err(error_response)
+                        }
+                    }
+                } else {
+                    match resp.json::<ErrorResponseBody>().await {
+                        Ok(body) => {
+                            let error_response = ErrorResponse { status, body };
+                            Err(error_response)
+                        }
+                        Err(err) => {
+                            let body = ErrorResponseBody {
+                                error: true,
+                                messages: vec![err.to_string()],
+                            };
+                            let error_response = ErrorResponse { status, body };
+                            Err(error_response)
+                        }
+                    }
+                }
+            }
+            Err(err) => {
+                let body = ErrorResponseBody {
+                    error: true,
+                    messages: vec![err.to_string()],
+                };
+                let error_response = ErrorResponse { status: 500, body };
+                Err(error_response)
+            }
         }
     }
 }
