+++ src/response/mystatus.rs
@@ -1,3 +1,4 @@
 pub mod mydeterms;
 pub mod mydrafts;
 pub mod myprocesses;
+pub mod myrequests;
